Factoring out a separate enrollment service. Adding test of one user with multiple subpopulations.
