+++ test/org/sagebionetworks/bridge/services/StudyServiceImplTest.java
@@ -19,46 +19,30 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import org.sagebionetworks.bridge.TestUserAdminHelper;
-import org.sagebionetworks.bridge.TestUserAdminHelper.TestUser;
 import org.sagebionetworks.bridge.TestUtils;
 import org.sagebionetworks.bridge.cache.CacheProvider;
 import org.sagebionetworks.bridge.dao.DirectoryDao;
-import org.sagebionetworks.bridge.dao.StudyConsentDao;
 import org.sagebionetworks.bridge.dao.SubpopulationDao;
-import org.sagebionetworks.bridge.dao.ParticipantOption.SharingScope;
 import org.sagebionetworks.bridge.dynamodb.DynamoStudy;
 import org.sagebionetworks.bridge.exceptions.EntityAlreadyExistsException;
 import org.sagebionetworks.bridge.exceptions.EntityNotFoundException;
 import org.sagebionetworks.bridge.exceptions.InvalidEntityException;
 import org.sagebionetworks.bridge.exceptions.UnauthorizedException;
-import org.sagebionetworks.bridge.models.accounts.ConsentStatus;
-import org.sagebionetworks.bridge.models.accounts.User;
 import org.sagebionetworks.bridge.models.studies.EmailTemplate;
 import org.sagebionetworks.bridge.models.studies.MimeType;
 import org.sagebionetworks.bridge.models.studies.PasswordPolicy;
 import org.sagebionetworks.bridge.models.studies.Study;
-import org.sagebionetworks.bridge.models.subpopulations.ConsentSignature;
 import org.sagebionetworks.bridge.models.subpopulations.StudyConsentView;
-import org.sagebionetworks.bridge.models.subpopulations.Subpopulation;
 import org.sagebionetworks.bridge.models.subpopulations.SubpopulationGuid;
-import org.sagebionetworks.bridge.redis.JedisOps;
-import org.sagebionetworks.bridge.redis.RedisKey;
 
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
-import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 
 @ContextConfiguration("classpath:test-context.xml")
 @RunWith(SpringJUnit4ClassRunner.class)
 public class StudyServiceImplTest {
-
-    private static final String NUM_PARTICIPANTS_KEY = RedisKey.NUM_OF_PARTICIPANTS.getRedisKey("test");
-    
-    @Resource
-    private JedisOps jedisOps;
 
     @Resource
     StudyServiceImpl studyService;
@@ -67,22 +51,10 @@ public class StudyServiceImplTest {
     StudyConsentServiceImpl studyConsentService;
     
     @Resource
-    StudyConsentDao studyConsentDao;
-    
-    @Resource
     DirectoryDao directoryDao;
     
     @Resource
     SubpopulationDao subpopDao;
-    
-    @Resource
-    SubpopulationService subpopService;
-    
-    @Resource
-    ConsentService userConsentService;
-    
-    @Resource
-    TestUserAdminHelper helper;
     
     private CacheProvider cache;
     
@@ -178,7 +150,7 @@ public class StudyServiceImplTest {
         // Verify that all the dependent stuff has been deleted as well:
         assertNull(directoryDao.getDirectoryForStudy(study));
         assertEquals(0, subpopDao.getSubpopulations(study.getStudyIdentifier(), false, true).size());
-        assertEquals(0, studyConsentDao.getConsents(SubpopulationGuid.create(study.getIdentifier())).size());
+        assertEquals(0, studyConsentService.getAllConsents(SubpopulationGuid.create(study.getIdentifier())).size());
         study = null;
     }
     
@@ -299,156 +271,5 @@ public class StudyServiceImplTest {
     @Test(expected = UnauthorizedException.class)
     public void cantDeleteApiStudy() {
         studyService.deleteStudy("api");
-    }
-    
-    @Test
-    public void enforcesStudyEnrollmentLimit() {
-        User user = new User();
-        user.setConsentStatuses(
-            Lists.newArrayList(new ConsentStatus("Consent", "guid", true, true, true)));
-        
-        try {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-            
-            Study study = TestUtils.getValidStudy(ConsentServiceImplTest.class);
-            study.setIdentifier("test");
-            study.setMaxNumOfParticipants(2);
-
-            // Set the cache so we avoid going to DynamoDB. We're testing the caching layer
-            // in the service test, we'll test the DAO in the DAO test.
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-
-            boolean limit = studyService.isStudyAtEnrollmentLimit(study);
-            assertFalse("No limit reached", limit);
-            studyService.incrementStudyEnrollment(study, user);
-            studyService.incrementStudyEnrollment(study, user);
-            assertTrue("Limit reached", studyService.isStudyAtEnrollmentLimit(study));
-            assertEquals("2", jedisOps.get(NUM_PARTICIPANTS_KEY));
-        } finally {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-        }
-    }
-    
-    @Test
-    public void studyEnrollmentNotIncrementedOnSubsequentConsents() {
-        User user = new User();
-        user.setConsentStatuses(
-            Lists.newArrayList(new ConsentStatus("Consent", "guid", true, true, true)));
-        
-        try {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-            jedisOps.setnx(NUM_PARTICIPANTS_KEY, "0");
-            
-            Study study = TestUtils.getValidStudy(ConsentServiceImplTest.class);
-            study.setIdentifier("test");
-            study.setMaxNumOfParticipants(2);
-
-            studyService.incrementStudyEnrollment(study, user);
-            assertFalse(studyService.isStudyAtEnrollmentLimit(study));
-            
-            // On subsequent consents, user is not added and enrollment is not increased
-            user.setConsentStatuses(Lists.newArrayList(
-                new ConsentStatus("Consent", "guid", true, true, true),
-                new ConsentStatus("Consent", "guid2", true, true, true)
-            ));
-            studyService.incrementStudyEnrollment(study, user);
-            studyService.incrementStudyEnrollment(study, user);
-            studyService.incrementStudyEnrollment(study, user);
-            assertFalse(studyService.isStudyAtEnrollmentLimit(study));
-            assertEquals("1", jedisOps.get(NUM_PARTICIPANTS_KEY));
-        } finally {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-        }
-    }
-    
-    @Test
-    public void decrementingStudyWorks() {
-        User user = new User();
-        user.setConsentStatuses(
-            Lists.newArrayList(new ConsentStatus("Consent", "guid", true, false, true)));
-        
-        try {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-            jedisOps.setnx(NUM_PARTICIPANTS_KEY, "2");
-            
-            Study study = TestUtils.getValidStudy(ConsentServiceImplTest.class);
-            study.setIdentifier("test");
-            study.setMaxNumOfParticipants(2);
-
-            studyService.decrementStudyEnrollment(study, user);
-            assertFalse(studyService.isStudyAtEnrollmentLimit(study));
-            
-            studyService.decrementStudyEnrollment(study, user);
-            studyService.decrementStudyEnrollment(study, user);
-            studyService.decrementStudyEnrollment(study, user);
-            studyService.decrementStudyEnrollment(study, user);
-            assertFalse(studyService.isStudyAtEnrollmentLimit(study));
-            assertEquals("0", jedisOps.get(NUM_PARTICIPANTS_KEY));
-        } finally {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-        }        
-    }
-    
-    @Test
-    public void studyEnrollmentNotDecrementedUntilLastWithdrawal() {
-        User user = new User();
-        user.setConsentStatuses(
-            Lists.newArrayList(new ConsentStatus("Consent", "guid", true, true, true)));
-        
-        try {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-            jedisOps.setnx(NUM_PARTICIPANTS_KEY, "2");
-            
-            Study study = TestUtils.getValidStudy(ConsentServiceImplTest.class);
-            study.setIdentifier("test");
-            study.setMaxNumOfParticipants(2);
-
-            // With a consent, this does not decrement
-            studyService.decrementStudyEnrollment(study, user);
-            assertEquals("2", jedisOps.get(NUM_PARTICIPANTS_KEY));
-            
-            // With one consent not signed, this will decrement.
-            user.setConsentStatuses(
-                    Lists.newArrayList(new ConsentStatus("Consent", "guid", true, false, true)));
-            studyService.decrementStudyEnrollment(study, user);
-            assertEquals("1", jedisOps.get(NUM_PARTICIPANTS_KEY));
-        } finally {
-            jedisOps.del(NUM_PARTICIPANTS_KEY);
-        }           
-    }
-    
-    @Test
-    public void getNumberOfParticipants() {
-        Study study = TestUtils.getValidStudy(ConsentServiceImplTest.class);
-        study = studyService.createStudy(study);
-        
-        Subpopulation subpop1 = Subpopulation.create();
-        subpop1.setName("Group 1");
-        subpopService.createSubpopulation(study, subpop1);
-        
-        Subpopulation subpop2 = Subpopulation.create();
-        subpop2.setName("Group 2");
-        subpopService.createSubpopulation(study, subpop2);
-        
-        TestUserAdminHelper.Builder builder = helper.getBuilder(StudyServiceImplTest.class).withStudy(study)
-                .withConsent(true);
-        
-        TestUser user1 = builder.withSubpopulation(subpop1).build();
-        TestUser user2 = builder.withSubpopulation(subpop1).build();
-        TestUser user3 = builder.withSubpopulation(subpop1).build();
-        // and user2 is also in the other subpop2, but will still be counted correctly.
-        userConsentService.consentToResearch(study, subpop2, user2.getUser(), 
-                new ConsentSignature.Builder().withBirthdate("1980-01-01").withName("Name").build(), SharingScope.NO_SHARING, false);
-        try {
-
-            long count = studyService.getNumberOfParticipants(study);
-            assertEquals(3, count);
-            
-        } finally {
-            helper.deleteUser(user1);
-            helper.deleteUser(user2);
-            helper.deleteUser(user3);
-            studyService.deleteStudy(study.getIdentifier());
-        }
     }
 }
